requirements.txt
spotipy
pandas
psycopg2-binary
python-dotenv
requests
dagster
dbt-core
dbt-snowflake
snowflake-connector-python
spotipy
pandas
psycopg2-binary
python-dotenv
requests
dagster
dbt-core
dbt-snowflake